+++ test/widget_test/country_test.dart
@@ -40,4 +40,37 @@ void main() => group('Country -', () {
       expect(translatedName, equals('United States'));
     });
   });
+  group('toLocalizedString', () {
+    testWidgets('should return localized country name if available', (
+      tester,
+    ) async {
+      final country = Country.fromJson(const {
+        'e164_cc': '1',
+        'iso2_cc': 'US',
+        'e164_sc': 1,
+        'geographic': true,
+        'level': 1,
+        'name': 'United States',
+        'example': '2012345678',
+        'display_name': 'United States (US) [1]',
+        'display_name_no_e164_cc': 'United States',
+        'e164_key': 'us',
+      });
+
+      await tester.pumpWidget(
+        createWidgetUnderTest(
+          locale: const Locale('en'),
+          builder: (context) =>
+              const Scaffold(key: _key, body: SizedBox.shrink()),
+        ),
+      );
+      await tester.pumpAndSettle();
+
+      expect(find.byKey(_key), findsOneWidget);
+
+      final context = tester.firstElement(find.byKey(_key));
+      final translatedName = country.toLocalizedString(context);
+      expect(translatedName, equals('United States'));
+    });
+  });
 });